Expose mechanism to set control bit to applications
As of this commit, UdpEchoClient can choose to SendControlPacket
or SendDataPacket. The former goes through the DDC code, while the
latter goes through the stock ns3 ipv4 routing code.

The next step is to enable the same for UdpEchoServer. Note that
there is currently only mechanism to do this for applications that
use the UDP socket, though the mechanism can be exposed to TCP
sockets in a very similar way.